refactor: move the work controller to v1alpha1
