+++ pkg/controllers/workv1alpha1/applied_work_syncer_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package workv1alpha1
 
 import (
 	"context"
