+++ pkg/controllers/workv1alpha1/apply_controller_helper_test.go
@@ -1,4 +1,4 @@
-package controllers
+package workv1alpha1
 
 import (
 	"context"
