+++ pkg/controllers/workv1alpha1/patch_util.go
@@ -1,4 +1,4 @@
-package controllers
+package workv1alpha1
 
 import (
 	"encoding/json"
